repositories-service-controller

+++ src/main/java/com/kiwi/entities/Basket.java
@@ -22,7 +22,9 @@ public class Basket {
     @GeneratedValue
     private Long id;
 
-    private Integer quantity;
+    private long quantity;
+
+
 
     /*
     @OneToMany
